% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{build_question_tag_matrix}
\alias{build_question_tag_matrix}
\title{Build a question tag matrix from cluster tags}
\usage{
build_question_tag_matrix(assignments, clusters)
}
\arguments{
\item{assignments}{Question-to-cluster assignments.}

\item{clusters}{Cluster tags with level and cluster ID.}
}
\value{
Tag matrix with \verb{tag_level_*} columns.
}
\description{
Build a question tag matrix from cluster tags
}
