% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence.R
\name{cosine_distance_to_centroid}
\alias{cosine_distance_to_centroid}
\title{Calculate cosine distance from a centroid}
\usage{
cosine_distance_to_centroid(mat, centroid)
}
\arguments{
\item{mat}{Numeric matrix with one observation per row.}

\item{centroid}{Numeric centroid vector.}
}
\value{
Numeric vector of cosine distances.
}
\description{
Calculate cosine distance from a centroid
}
\keyword{internal}
