% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama-provider.R
\name{ollama_model_provider}
\alias{ollama_model_provider}
\title{Create an optional Ollama model-provider adapter}
\usage{
ollama_model_provider(config = ollama_config())
}
\arguments{
\item{config}{Output from \code{\link[=ollama_config]{ollama_config()}}.}
}
\value{
A \code{model_provider}.
}
\description{
Create an optional Ollama model-provider adapter
}
