% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging-store.R
\name{tag_store_exists}
\alias{tag_store_exists}
\title{Test whether a tagging store contains state}
\usage{
tag_store_exists(store)
}
\arguments{
\item{store}{A \code{tag_store}.}
}
\value{
A logical scalar.
}
\description{
Test whether a tagging store contains state
}
