% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging-store.R
\name{tag_store_load}
\alias{tag_store_load}
\title{Load state from a tagging store}
\usage{
tag_store_load(store)
}
\arguments{
\item{store}{A \code{tag_store}.}
}
\value{
A validated \code{tag_state}.
}
\description{
Load state from a tagging store
}
